# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: F1 Live Telecronaca API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /races:
    get:
      operationId: listRaces
      tags: [races]
      summary: List all races
      responses:
        "200":
          description: List of races
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Race"
    post:
      operationId: createRace
      tags: [races]
      summary: Create a race (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Race"
...
  /races/{id}:
    get:
      operationId: getRace
      tags: [races]
      summary: Get race by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Race detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Race"
    patch:
      operationId: updateRace
      tags: [races]
      summary: Update a race (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Race"
    delete:
      operationId: deleteRace
      tags: [races]
      summary: Delete a race (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /races/{id}/summary:
    get:
      operationId: getRaceSummary
      tags: [races]
      summary: Get race stats summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Race summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaceSummary"
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /races/{id}/tires:
    get:
      operationId: listTireStrategies
      tags: [tires]
      summary: List tire strategies for a race
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tire strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TireStrategy"
    post:
      operationId: addTireStrategy
      tags: [tires]
      summary: Add tire stint (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TireStrategyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TireStrategy"
components:
  schemas:
...
    Race:
      type: object
      required: [id, name, circuit, country, round, season, status, raceDate]
      properties:
        id:
          type: integer
        name:
          type: string
        circuit:
          type: string
        country:
          type: string
        round:
          type: integer
        season:
          type: integer
        status:
          type: string
          enum: [upcoming, practice, qualifying, race, finished]
        raceDate:
          type: string
        currentLap:
          type: ["integer", "null"]
        totalLaps:
          type: ["integer", "null"]
        weatherCondition:
          type: ["string", "null"]
        trackTemp:
          type: ["number", "null"]
        airTemp:
          type: ["number", "null"]
        safetyCarDeployed:
          type: boolean
        vscDeployed:
          type: boolean
    RaceInput:
      type: object
      required: [name, circuit, country, round, season, raceDate]
      properties:
        name:
          type: string
        circuit:
          type: string
        country:
          type: string
        round:
          type: integer
        season:
          type: integer
        raceDate:
          type: string
        totalLaps:
          type: ["integer", "null"]
        weatherCondition:
          type: ["string", "null"]
        trackTemp:
          type: ["number", "null"]
        airTemp:
          type: ["number", "null"]
    RaceUpdate:
      type: object
      properties:
        name:
          type: string
        circuit:
          type: string
        country:
          type: string
        round:
          type: integer
        raceDate:
          type: string
        status:
          type: string
          enum: [upcoming, practice, qualifying, race, finished]
        currentLap:
          type: ["integer", "null"]
        totalLaps:
          type: ["integer", "null"]
        weatherCondition:
          type: ["string", "null"]
        trackTemp:
          type: ["number", "null"]
        airTemp:
          type: ["number", "null"]
        safetyCarDeployed:
          type: boolean
        vscDeployed:
          type: boolean
...
    RaceSummary:
      type: object
      required: [raceId, totalDrivers, finishedDrivers, dnfCount]
      properties:
        raceId:
          type: integer
        totalDrivers:
          type: integer
        finishedDrivers:
          type: integer
        dnfCount:
          type: integer
        fastestLap:
          type: ["string", "null"]
        fastestLapDriver:
          type: ["string", "null"]
        totalLapsCompleted:
          type: integer
        safetyCarPeriods:
          type: integer
...
    Team:
      type: object
      required: [id, name, color, nationality, principal]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        nationality:
          type: string
        principal:
          type: string
        chassis:
          type: ["string", "null"]
        powerUnit:
          type: ["string", "null"]
    TeamInput:
      type: object
      required: [name, color, nationality, principal]
      properties:
        name:
          type: string
        color:
          type: string
        nationality:
          type: string
        principal:
          type: string
        chassis:
          type: ["string", "null"]
        powerUnit:
          type: ["string", "null"]
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        nationality:
          type: string
        principal:
          type: string
        chassis:
          type: ["string", "null"]
        powerUnit:
          type: ["string", "null"]
...
    TireStrategy:
      type: object
      required: [id, raceId, driverId, compound, startLap, driverName]
      properties:
        id:
          type: integer
        raceId:
          type: integer
        driverId:
          type: integer
        driverName:
          type: string
        compound:
          type: string
          enum: [soft, medium, hard, intermediate, wet]
        startLap:
          type: integer
        endLap:
          type: ["integer", "null"]
    TireStrategyInput:
      type: object
      required: [driverId, compound, startLap]
      properties:
        driverId:
          type: integer
        compound:
          type: string
          enum: [soft, medium, hard, intermediate, wet]
        startLap:
          type: integer
        endLap:
          type: ["integer", "null"]