# lib/api-zod/src/generated/types/raceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * F1 Live Telecronaca API
 * OpenAPI spec version: 0.1.0
 */

export interface RaceInput {
  name: string;
  circuit: string;
  country: string;
  round: number;
  season: number;
  raceDate: string;
  /** @nullable */
  totalLaps?: number | null;
  /** @nullable */
  weatherCondition?: string | null;
  /** @nullable */
  trackTemp?: number | null;
  /** @nullable */
  airTemp?: number | null;
}