# lib/api-zod/src/generated/types/raceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * F1 Live Telecronaca API
 * OpenAPI spec version: 0.1.0
 */
import type { RaceUpdateStatus } from './raceUpdateStatus';

export interface RaceUpdate {
  name?: string;
  circuit?: string;
  country?: string;
  round?: number;
  raceDate?: string;
  status?: RaceUpdateStatus;
  /** @nullable */
  currentLap?: number | null;
  /** @nullable */
  totalLaps?: number | null;
  /** @nullable */
  weatherCondition?: string | null;
  /** @nullable */
  trackTemp?: number | null;
  /** @nullable */
  airTemp?: number | null;
  safetyCarDeployed?: boolean;
  vscDeployed?: boolean;
}