# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * F1 Live Telecronaca API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamUpdate {
  name?: string;
  color?: string;
  nationality?: string;
  principal?: string;
  /** @nullable */
  chassis?: string | null;
  /** @nullable */
  powerUnit?: string | null;
}