# lib/api-zod/src/generated/types/tireStrategyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * F1 Live Telecronaca API
 * OpenAPI spec version: 0.1.0
 */
import type { TireStrategyInputCompound } from './tireStrategyInputCompound';

export interface TireStrategyInput {
  driverId: number;
  compound: TireStrategyInputCompound;
  startLap: number;
  /** @nullable */
  endLap?: number | null;
}